ITSVA_BOT.02/Gabi_Bot/requirements.txt
pyTelegramBotAPI
mysql-connector-python
python-dotenv